+++ lib/application.ex
@@ -4,10 +4,6 @@ defmodule BridgeEx.Application do
   use Application
 
   def start(_type, _args) do
-    if Application.get_env(:bridge_ex, :auth0_enabled) do
-      {:ok, _} = Application.ensure_all_started(:prima_auth0_ex)
-    end
-
     children = []
     Supervisor.start_link(children, strategy: :one_for_one)
   end