+++ lib/graphql.ex
@@ -138,28 +138,14 @@ defmodule BridgeEx.Graphql do
 
       if @audience && @auth0_enabled do
         defp with_authorization_headers(headers) do
-          with {:ok, authorization_headers} <- get_authorization_headers_for(@audience) do
+          with {:ok, authorization_headers} <-
+                 Auth0AuthorizationProvider.authorization_headers(@audience) do
             {:ok, Enum.into(authorization_headers, headers)}
           end
         end
       else
         defp with_authorization_headers(headers), do: {:ok, headers}
       end
-
-      defp get_authorization_headers_for(audience) do
-        if auth0_enabled_for_app(),
-          do: Auth0AuthorizationProvider.authorization_headers(@audience),
-          else: report_auth0_disabled_error()
-      end
-
-      defp report_auth0_disabled_error do
-        Logger.error("Auth0 is not enabled for this application!
-        To enable it set `config :bridge_ex, auth0_enabled: true` in your config.")
-
-        {:error, "Auth0 not enabled in application"}
-      end
-
-      defp auth0_enabled_for_app, do: Application.get_env(:bridge_ex, :auth0_enabled, false)
     end
   end
 end