+++ test/config_validation_test.exs
@@ -11,7 +11,7 @@ defmodule BridgeEx.ConfigValidationTest do
   # This test doesn't play nice with others and apparently needs to be on his own, otherwise flakiness ensues.
   # We investigated this issue for quite some time but this is the only solution we could come up with :(
   test "macro won't expand if auth0 is enabled but no audience is set", %{bypass: bypass} do
-    assert_raise CompileError, fn ->
+    assert_raise RuntimeError, fn ->
       defmodule TestBridgeWithAuth0EnabledButNoAudience do
         use BridgeEx.Graphql,
           endpoint: "http://localhost:#{bypass.port}/graphql",