+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Next]
 
+### Changed
+
+- Bump minimum Elixir version to 1.13
+
 ## [1.0.0] - 2022-02-18
 
 ### Changed